planventure-api/requirements.txt
# Core dependencies
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0

# Authentication and security
flask-jwt-extended==4.5.2
bcrypt==4.0.1
email-validator==2.0.0

# Testing and development
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Production
gunicorn==21.2.0